meal_planner/requirements.txt
Flask
mysql-connector-python
flask-bcrypt 